tripit-to-flightdiary/pyproject.toml
[tool.poetry]
name = "tripit-to-flightdiary"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
requests-oauthlib = "^1.3.0"
beautifulsoup4 = "^4.9.3"
html5lib = "^1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
